firecloud - adding,removing,editing courses implemented

+++ src/app/courses-list/courses-list.component.ts
@@ -2,39 +2,42 @@ import { Component, OnInit } from '@angular/core';
 import { Course } from '../interfaces/course';
 import { CoursesService } from '../courses.service';
 import { FilterInterface } from '../interfaces/filterInterface';
+import { FilterService } from '../filter.service';
+import { FilterPipe } from '../pipes/filter-pipe'
+import { AuthService } from '../auth.service';
 
 @Component({
-  selector: 'app-courses-list',
-  templateUrl: './courses-list.component.html',
-  styleUrls: ['./courses-list.component.css']
+	selector: 'app-courses-list',
+	templateUrl: './courses-list.component.html',
+	styleUrls: ['./courses-list.component.css']
 })
 export class CoursesListComponent implements OnInit {
 
-  coursesList: Array<Course>;
-  filterCriteria: FilterInterface;
+	coursesList: Array<Course>;
+	filterCriteria: FilterInterface;
 
-  constructor(private courseService: CoursesService) { 
-    this.filterCriteria = {
-      ectsValues: [],
-      rateValues: [],
-      semesterValues: [],
-      textValue: ""
-    }    
-  }
+	constructor(
+		private courseService: CoursesService,
+		private filters: FilterService,
+		private authService: AuthService
+	) {
+		this.filterCriteria = new FilterInterface()
+	}
 
-  ngOnInit() {
-    this.getCourses();
-  }
+	ngOnInit() {
+		this.courseService.getCourses().subscribe(courses => {
+			this.coursesList = courses
+		});
 
-  getCourses() : void {
-    this.courseService.getCourses().subscribe(courses => this.coursesList = courses);
-  }
+		this.filters.filterCriteria$.subscribe(fr => {
+			this.filterCriteria = fr
+		})
+	}
 
-  onDeleteSignal(crs : Course) : void{
-    this.courseService.deleteCourse(crs);
-  }
-
-  filterChanged(filterCriteria : FilterInterface) : void{
-    this.filterCriteria = filterCriteria;
-  }
+	onDeleteSignal(crs: Course): void {
+		if(this.authService.isAdmin()){
+			this.coursesList = this.coursesList.filter(c => c.id !== crs.id)
+			this.courseService.deleteCourse(crs);
+		}
+	}
 }
